quantum-ai-trading-system/production/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
cmdstanpy==1.3.0
colorlog==6.10.1
contourpy==1.3.3
cramjam==2.11.0
curl_cffi==0.13.0
cycler==0.12.1
darts==0.38.0
fastapi==0.121.0
fastparquet==2024.11.0
filelock==3.20.0
fonttools==4.60.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.2.0
holidays==0.84
huggingface-hub==0.36.0
idna==3.11
importlib_resources==6.5.2
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lightgbm==4.6.0
lightning-utilities==0.15.2
llvmlite==0.45.1
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.7.0
multitasking==0.0.12
narwhals==2.11.0
nest-asyncio==1.6.0
networkx==3.5
nfoursid==1.0.2
nltk==3.9.2
numba==0.62.1
numpy==2.3.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
optuna==4.5.0
packaging==25.0
pandas==2.3.3
patsy==1.0.2
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.0
plotly==6.4.0
pluggy==1.6.0
propcache==0.4.1
prophet==1.2.1
protobuf==6.33.1
psutil==7.1.3
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyod==2.0.5
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-lightning==2.5.2
pytz==2025.2
PyWavelets==1.9.0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
shap==0.50.0
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
stanio==0.5.1
starlette==0.49.3
statsmodels==0.14.5
sympy==1.14.0
tensorboardX==2.6.4
textblob==0.19.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
torchmetrics==1.8.2
tqdm==4.67.1
transformers==4.57.1
triton==3.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
vaderSentiment==3.3.2
websockets==15.0.1
xarray==2025.10.1
xgboost==3.1.1
yarl==1.22.0
yfinance==0.2.66
streamlit==1.40.0
